requirements.txt
line-bot-sdk
flask
yfinance
gunicorn